flask_crud_app/requirements.txt
flask
psycopg2-binary
tenacity